Refactor animated mesh rendering by integrating Renderer3D; remove legacy AnimatedMeshRenderSystem files and update related references in Sandbox3D.
