+++ OloEngine/src/CMakeLists.txt
@@ -3,14 +3,12 @@
 		"OloEnginePCH.cpp"
 		"OloEnginePCH.h"
 
-		"OloEngine/Animation/Skeleton.h"
-		"OloEngine/Animation/AnimatedMeshComponents.h"
-		"OloEngine/Animation/AnimatedMeshRenderSystem.h"
-		"OloEngine/Animation/AnimatedMeshRenderSystem.cpp"
-		"OloEngine/Animation/AnimationSystem.h"
-		"OloEngine/Animation/AnimationSystem.cpp"
-		"OloEngine/Animation/AnimationClip.h"
-		"OloEngine/Animation/AnimationClip.cpp"
+	"OloEngine/Animation/Skeleton.h"
+	"OloEngine/Animation/AnimatedMeshComponents.h"
+	"OloEngine/Animation/AnimationSystem.h"
+	"OloEngine/Animation/AnimationSystem.cpp"
+	"OloEngine/Animation/AnimationClip.h"
+	"OloEngine/Animation/AnimationClip.cpp"
 		
 		"OloEngine/Audio/AudioEngine.cpp"
 		"OloEngine/Audio/AudioEngine.h"
